--- apps/main/models/__init__.py
@@ -1,21 +1,0 @@
-from .notification import Notification
-from .business import (
-    TermsAndConditions,
-    PrivacyPolicy,
-    Contact,
-    AuditLogConfig,
-    SocialMediaLink,
-    FAQ,
-    Report,
-)
-
-__all__ = [
-    "Notification",
-    "TermsAndConditions",
-    "PrivacyPolicy",
-    "Contact",
-    "AuditLogConfig",
-    "SocialMediaLink",
-    "FAQ",
-    "Report",
-]
